backend_main/requirements.txt
aioredis==2.0.1
amqp==5.3.1
annoy==1.17.3
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.12.0
cron_descriptor==2.0.5
cryptography==45.0.6
daphne==4.2.0
Django==5.1.8
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-fernet-fields==0.6
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
incremental==24.7.2
iniconfig==2.3.0
isort==6.0.1
joblib==1.5.2
kombu==5.5.4
lazy_loader==0.4
mongoengine==0.29.1
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg-pool==3.2.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pymongo==4.14.1
pyOpenSSL==25.1.0
pytest==9.0.1
pytest-cov==7.0.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1.1
